Content_Generation/requirements.txt
# RAG + OpenDebateEvidence(OpenCaselist) + ColBERTv2
#
# Usage (inside your conda env "content"):
#   python -m pip install -U pip
#   python -m pip install -r requirements_rag_colbert.txt
#
# Notes:
# - Install PyTorch separately to match your CUDA version (or CPU-only).
#   See: https://pytorch.org/get-started/locally/
# - If `colbert-ai` pip install fails on Windows, use the git fallback shown below.

# ---- Core data loading (HuggingFace datasets) ----
datasets>=2.16.1
huggingface_hub>=0.20.3
pyarrow>=14.0.2
fsspec>=2023.10.0
requests>=2.31.0

# ---- Common utilities ----
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
orjson>=3.9.0

# ---- LLM / tokenizer deps (also used by ColBERT) ----
transformers>=4.38.0
accelerate>=0.26.0
sentencepiece>=0.1.99
regex>=2023.12.25
ftfy>=6.1.1
ujson>=5.9.0
ninja>=1.11.1

# ---- Retrieval index deps ----
# Prefer GPU if you have CUDA: use faiss-gpu (conda usually), else faiss-cpu via pip.
faiss-cpu>=1.7.4

# ---- ColBERTv2 ----
# Primary: pip package
colbert-ai>=0.2.20

# Fallback (uncomment if needed): install from GitHub
# git+https://github.com/stanford-futuredata/ColBERT.git
